Embryo_Tracker: changed Hungarian classes to use double[] instead of float[].
Changed the Hungarian classes written by Gary Baker to use double and double[] internally instead of float[] and float, so I could use Jama class easily.

+++ src-plugins/Embryo_Tracker/fiji/plugin/nperry/hungarian/AssignmentAlgorithm.java
@@ -8,6 +8,6 @@ package fiji.plugin.nperry.hungarian;
 public interface AssignmentAlgorithm {
 
 
-    int[][] computeAssignments(float[][] costMatrix);
+    int[][] computeAssignments(double[][] costMatrix);
 
 }
